Added a helper file for FilterBar component.
